requirements.txt
certifi==2019.9.11
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
numpy==1.17.2
openpyxl==3.0.0
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
urllib3==1.25.6
python-telegram-bot==12.2.0
